ISSUE-21, 26: Multi-enemies, multi-waves spawning across different frames now supported

+++ src/voogasalad/gameengine/engine/sprites/JavaFXSpriteManager.java
@@ -24,7 +24,7 @@ public class JavaFXSpriteManager implements SpriteManager {
     }
 
     @Override
-    public void makeSpriteFromPrototype(int xPos, int yPos, int prototypeId) throws GameEngineException {
+    public void makeSpriteFromPrototype(double xPos, double yPos, int prototypeId) throws GameEngineException {
         myOnScreenSprites.add(mySpritePrototypes.get(prototypeId).makeClone(xPos, yPos, mySpriteIDGenerator++));
     }
