+++ src/voogasalad/gameengine/engine/sprites/Sprite.java
@@ -1,12 +1,11 @@
 package voogasalad.gameengine.engine.sprites;
 
 import voogasalad.gameengine.engine.exceptions.GameEngineException;
-import voogasalad.gameengine.engine.sprites.strategies.health.HealthStrategy;
 
 public interface Sprite {
-    Sprite makeClone(int x, int y, int spriteId) throws GameEngineException;
-    int getX();
-    int getY();
+    Sprite makeClone(double x, double y, int spriteId) throws GameEngineException;
+    double getX();
+    double getY();
     int getId();
     int getHealth();
 }
