+++ src/voogasalad/gameengine/engine/sprites/SpriteManager.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface SpriteManager {
     void addSpritePrototype(int prototypeId, Sprite sprite);
-    void makeSpriteFromPrototype(int xPos, int yPos, int prototypeId) throws GameEngineException;
+    void makeSpriteFromPrototype(double xPos, double yPos, int prototypeId) throws GameEngineException;
     List<Sprite> getOnScreenSprites();
 }
